pyproject.toml
[project]
name = "discordbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py>=2.4.0",
    "pygithub>=2.5.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
